refactor(patina): extract scope-variable parsers into visitor_scope module
visitor.rs was already over the 350-line limit and grew by 77 lines from
the slot-scope and binding-normalization additions, tripping the
source-file-length guard (706 -> 783 lines).

Move parse_v_for_variables, parse_slot_scope_variables, and their
private helpers into a new visitor_scope module. Re-export the two
public functions from visitor so existing import paths
(crate::visitor::parse_v_for_variables) are unchanged.
